backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
google-generativeai==0.8.3
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
python-multipart==0.0.20
